refactor: rename queue module
